refactor(router): move location specs into their own folder
