+++ modules/angular2/test/router/location/location_spec.ts
